+++ docs/_includes/eleventyComponents.js
@@ -6,17 +6,14 @@
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
 
-import acutis from "acutis-lang";
 import site from "../_data/site.js";
-
-let Typescheme = acutis.Typescheme;
 
 export function Footer(props) {
   let year = props.year ? props.year : new Date().getFullYear();
   let name = props.link
     ? `<a href="${props.link}">${props.name}</a>`
     : props.name;
-  return Promise.resolve(`
+  return `
     <footer class="footer">
       <p>
         Published in ${year} by ${name}.
@@ -24,24 +21,22 @@ export function Footer(props) {
       <p class="footer__license">
         <a href="${props.siteUrl}/license/">View the license</a>.
       </p>
-    </footer>`);
+    </footer>`;
 }
-Footer.interface = Typescheme.make([
-  ["year", Typescheme.nullable(Typescheme.string())],
-  ["link", Typescheme.nullable(Typescheme.string())],
-  ["name", Typescheme.string()],
-  ["siteUrl", Typescheme.string()],
-]);
+Footer.interface = {
+  year: ["nullable", "string"],
+  link: ["nullable", "string"],
+  name: "string",
+  siteUrl: "string",
+};
 
 export function Link({ path, page, children }) {
   let current = path === page.url ? "true" : "false";
   let href = site.url + path;
-  return Promise.resolve(
-    `<a href="${href}" aria-current="${current}">${children}</a>`,
-  );
+  return `<a href="${href}" aria-current="${current}">${children}</a>`;
 }
-Link.interface = Typescheme.make([
-  ["path", Typescheme.string()],
-  ["page", Typescheme.record([["url", Typescheme.string()]])],
-  ["children", Typescheme.string()],
-]);
+Link.interface = {
+  path: "string",
+  page: { url: "string" },
+  children: "string",
+};