+++ test/data-test.t/run.js
@@ -1,58 +1,50 @@
 import fs from "node:fs/promises";
 import acutis from "#main";
-let { Compile, Component, Typescheme, Utils } = acutis;
+let { Compile, Component, Utils } = acutis;
 
 let filepath = process.argv[2];
-let Ty = Typescheme;
-
-let components = Compile.components([
-  Component.func(
-    "Comp",
-    Ty.make([
-      [
-        "dict",
-        Ty.record([
-          ["a", Ty.int()],
-          ["b", Ty.string()],
-        ]),
-      ],
-      ["arr", Ty.list(Ty.string())],
-      ["arrEmpty", Ty.list(Ty.string())],
-      ["some", Ty.nullable(Ty.string())],
-      ["none", Ty.nullable(Ty.string())],
-      ["t", Ty.boolean()],
-      ["f", Ty.boolean()],
-      ["unknown", Ty.unknown()],
-    ]),
-    (props) => JSON.stringify(props, null, 2),
-  ),
-  Component.func(
-    "UnknownComp",
-    Ty.make([
-      // The marshaled representation of these values is not deterministic.
-      // Uncomment to test them but don't commit the output.
-      /*
-      ["dict", Ty.unknown()],
-      ["arr", Ty.unknown()],
-      ["arrEmpty", Ty.unknown()],
-      ["some", Ty.unknown()],
-      ["none", Ty.unknown()],
-      ["t", Ty.unknown()],
-      ["f", Ty.unknown()],
-      */
-      ["unknown", Ty.unknown()],
-    ]),
-    (props) => JSON.stringify(props, null, 2),
-  ),
-]);
 
 Promise.all([fs.readFile("data.json"), fs.readFile(filepath)])
-  .then(async ([dataSrc, src]) => {
+  .then(([dataSrc, src]) => {
+    let components = Compile.components([
+      Component.func(
+        "Comp",
+        {
+          dict: { a: "int", b: "string" },
+          arr: ["list", "string"],
+          arrEmpty: ["list", "string"],
+          some: ["nullable", "string"],
+          none: ["nullable", "string"],
+          t: ["enum", false, true],
+          f: ["enum", false, true],
+          unknown: "_",
+        },
+        (props) => JSON.stringify(props, null, 2),
+      ),
+      Component.func(
+        "UnknownComp",
+        {
+          // The marshaled representation of these values is not deterministic.
+          // Uncomment to test them but don't commit the output.
+          /*
+          dict: "_",
+          arr: "_",
+          arrEmpty: "_",
+          some: "_",
+          none: "_",
+          t: "_",
+          f: "_",
+          */
+          unknown: "_",
+        },
+        (props) => JSON.stringify(props, null, 2),
+      ),
+    ]);
     let data = JSON.parse(dataSrc);
     let template = Compile.uint8Array(filepath, components, src);
-    let result = await Compile.render(template, data);
-    process.stdout.write(result);
+    return Compile.render(template, data);
   })
+  .then((result) => process.stdout.write(result))
   .catch((e) => {
     if (Utils.isError(e)) {
       console.error(Utils.getError(e));